listevents/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
cssselect2==0.4.1
Django==3.2.11
django-countries==7.2.1
django-crispy-forms==1.13.0
django-filter==21.1
django-js-asset==1.2.2
django-mptt==0.13.4
django-phonenumber-field==6.0.0
django-polymorphic==3.0.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
easy-thumbnails==2.8.1
lxml==4.7.1
phonenumbers==8.12.41
Pillow==9.0.0
psycopg2-binary==2.9.3
pytz==2021.3
reportlab==3.6.6
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
Unidecode==1.1.2
webencodings==0.5.1
